# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rota Certa Passagens - API do site público
  version: 0.1.0
  description: Contrato da API do site público e do painel administrativo.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rc_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
    cronToken:
      type: http
      scheme: bearer
      description: NOTIFICATIONS_CRON_TOKEN, for unattended scheduler calls with no session cookie.
...
paths:
...
  /api/admin/finance/obligations:
    get:
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      summary: Lista contas a pagar / despesas, com is_overdue calculado no servidor (status IN (open,partial) AND due_date < hoje) sem sobrescrever o status armazenado
      parameters:
        - { name: status, in: query, required: false, schema: { type: string, enum: [open, partial, paid, canceled, reversed] } }
        - { name: kind, in: query, required: false, schema: { type: string, enum: [direct_cost, operating_expense] } }
        - { name: dueBefore, in: query, required: false, schema: { type: string, format: date } }
        - { name: counterpartyId, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: currency, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: Lista de obrigações }
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Cria obrigação manual (custo direto ou despesa operacional); categoria precisa ter o mesmo kind
      responses:
        "201": { description: "{ id }" }
        "422": { description: category_kind_mismatch ou vínculo inválido }
  /api/admin/finance/obligations/{id}:
    patch:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Atualiza categoria/centro de custo/vencimento/conta/observações — bloqueado se a obrigação estiver cancelada ou estornada
      responses:
        "200": { description: "{ ok: true }" }
        "409": { description: obligation_terminal }
  /api/admin/finance/obligations/{id}/cancel:
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Cancela uma obrigação em aberto (sem nenhum pagamento) com motivo auditado — nunca DELETE
      responses:
        "200": { description: "{ ok: true }" }
        "409": { description: obligation_has_payments_or_terminal }
  /api/admin/finance/obligations/{id}/payments:
    get:
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      summary: Lista os pagamentos (e estornos) de uma obrigação, em ordem cronológica
      responses:
        "200": { description: Lista de pagamentos }
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Registra pagamento parcial ou total; recalcula o status da obrigação (open/partial/paid) no servidor
      responses:
        "201": { description: "{ id }" }
        "422": { description: payment_currency_must_match_obligation }
        "409": { description: obligation_not_payable }
  /api/admin/finance/obligation-payments/{id}/reverse:
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Estorna um pagamento como um novo lançamento negativo vinculado (reversal_of) — nunca apaga o pagamento original; no máximo um estorno por pagamento
      responses:
        "201": { description: "{ id }" }
        "409": { description: payment_already_reversed ou cannot_reverse_a_reversal }
...